Solved little merge conflicts introduced above original merge commit

+++ src/cpp/rtps/participant/RTPSParticipantImpl.cpp
@@ -107,7 +107,8 @@ RTPSParticipantImpl::RTPSParticipantImpl(const RTPSParticipantAttributes& PParam
 
     /// Creation of metatraffic locator and receiver resources
     uint32_t metatraffic_multicast_port = m_att.port.getMulticastPort(m_att.builtin.domainId);
-    uint32_t metatraffic_unicast_port = m_att.port.getUnicastPort(m_att.builtin.domainId, m_att.participantID);
+    uint32_t metatraffic_unicast_port = m_att.port.getUnicastPort(m_att.builtin.domainId,
+        static_cast<uint32_t>(m_att.participantID));
 
     // Workaround TCP discovery issues when register
     switch (PParam.builtin.discoveryProtocol)
